requirements.txt
appdirs==1.4.4
asgiref==3.8.1
atomicwrites==1.4.0
backoff==2.2.1
bcrypt==4.2.1
build==1.2.2.post1
chroma-hnswlib==0.7.6
chromadb==0.5.21
coloredlogs==15.0.1
dataclasses-json==0.6.7
Deprecated==1.2.15
docstring_parser==0.16
durationpy==0.9
fastapi==0.115.6
filetype==1.2.0
flatbuffers==24.3.25
git-filter-repo==2.47.0
google-ai-generativelanguage==0.6.10
google-api-core==2.23.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.73.0
google-cloud-bigquery==3.27.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.13.1
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-generativeai==0.8.3
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grpc-google-iam-v1==0.13.1
grpcio==1.68.1
grpcio-status==1.68.1
httplib2==0.22.0
httptools==0.6.4
httpx-sse==0.4.0
huggingface-hub==0.26.3
humanfriendly==10.0
importlib_resources==6.4.5
jiter==0.8.0
jsonpointer==2.1
kubernetes==31.0.0
langchain==0.3.9
langchain-chroma==0.1.4
langchain-community==0.3.9
langchain-core==0.3.21
langchain-google-genai==2.0.6
langchain-google-vertexai==2.0.8
langchain-openai==0.2.11
langchain-text-splitters==0.3.2
langserve==0.3.0
langsmith==0.1.147
marshmallow==3.23.1
mkl-service==2.4.0
mmh3==5.0.1
monotonic==1.6
oauthlib==3.2.2
onnxruntime==1.20.1
openai==1.56.1
opencv-python==4.10.0.84
opentelemetry-api==1.28.2
opentelemetry-exporter-otlp-proto-common==1.28.2
opentelemetry-exporter-otlp-proto-grpc==1.28.2
opentelemetry-instrumentation==0.49b2
opentelemetry-instrumentation-asgi==0.49b2
opentelemetry-instrumentation-fastapi==0.49b2
opentelemetry-proto==1.28.2
opentelemetry-sdk==1.28.2
opentelemetry-semantic-conventions==0.49b2
opentelemetry-util-http==0.49b2
orjson==3.10.12
posthog==3.7.4
proto-plus==1.25.0
protobuf==4.25.5
pyasn1-modules==0.2.8
pycurl==7.45.2
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
pyls-spyder==0.4.0
PyPika==0.48.9
pyproject_hooks==1.2.0
PyQt5==5.15.10
PyQtWebEngine==5.15.6
pyreadline3==3.5.4
python-docx==1.1.2
pywin32==305.1
requests-oauthlib==2.0.0
rsa==4.9
safetensors==0.4.5
setuptools==69.5.1
shapely==2.0.6
shellingham==1.5.4
starlette==0.41.3
sympy==1.13.1
tenacity==8.5.0
tiktoken==0.8.0
tokenizers==0.21.0
torch==2.5.1
torchvision==0.20.1
transformers==4.47.1
typer==0.15.0
typing-inspect==0.9.0
typing_extensions==4.12.2
ultralytics==8.3.41
ultralytics-thop==2.0.12
uritemplate==4.1.1
uvicorn==0.32.1
watchfiles==1.0.0
websockets==14.1
wheel==0.43.0
